requirements.txt
streamlit
PyPDF2
langchain
langchain-google-genai
google-generativeai
faiss-cpu
python-dotenv
speechrecognition
pyttsx3
langchain_core
python-dotenv
langchain_community
langserve
fastapi
uvicorn
sse_starlette
bs4
pypdf
chromadb
faiss-cpu

cassio
beautifulsoup4
langchain-groq

langchainhub
sentence_transformers

langchain-objectbox